Move this file too.
